+++ lib/flutter_screen_adapter.dart
@@ -1,5 +1,3 @@
 library flutter_screen_adapter;
 
 export 'src/media_query_builder.dart';
-export 'src/screen_adapter_binding.dart';
-export 'src/screen_util.dart';